Split up the db.js file into separate model files
