+++ backend/db/models/Item.js
@@ -1,0 +1,42 @@
+const sequelize = require('../db.js');
+const Item = require('./Item.js');
+
+const Item = sequelize.define('Item', {
+  id: {
+    type: DataTypes.INTEGER,
+    primaryKey: true,
+    autoIncrement: true
+  },
+  name: DataTypes.STRING,
+  description: DataTypes.STRING,
+  // Right now we just will be able to support a single picture. Make a sepparate image table to support multiple pictures?
+  imageUrl: DataTypes.STRING,
+  // Have the item location be based off the users current signed in location but they can change it if they want to
+  location: DataTypes.INTEGER,
+  claimed: {
+    type: DataTypes.BOOLEAN,
+    defaultValue: false
+  },
+  // We might also do a separate category table and then have this be a foreign key
+  category: DataTypes.STRING,
+  donorId: DataTypes.INTEGER
+},
+{
+  indexes: [
+    {
+      name: 'itemDonorIndex',
+      using: 'HASH',
+      fields: ['donorId']
+    }
+  ]
+}
+);
+
+// Setting up associations (Relationships)
+// Each donor (User) can have many donations (Items)
+User.hasMany(Item, {foreignKey: 'donorId'});
+Item.belongsTo(User);
+
+
+
+module.exports = Item;
